Move PaletteMap implementation in PaletteMap.cpp
